requirements.txt
click==7.1.2
colorama==0.4.3
Flask==1.1.2
Flask-QRcode==3.0.0
future==0.18.2
install==1.3.3
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
protobuf==3.13.0
psycopg2==2.8.6
pyserial==3.4
PyYAML==5.3.1
qrcode==6.1
six==1.15.0
Werkzeug==1.0.1
